CONSOLE-5196: Consolidate alertmanager receiver specs into single serial file
All four receiver specs (webhook, email, slack, pagerduty) mutate the
same cluster-wide alertmanager-main secret. With WORKERS=2, separate
spec files run in parallel, so one test's resetAlertmanagerConfig wipes
the receiver another test just created. Consolidating into a single file
with serial mode ensures they never overlap.
